hw2.R
I think the author is trying to portray the mitigation of vaccine for the disease brought a positive effect on the decrease of cases. Especially, the emphasis is on the policy for possible weak society including mothers, children and high-risk communities for the vaccination that brought a decrease of cases in different types of disease. Along the timeline of disease, the discovery of vaccine is clearly portrayed to show that the cases are decreased for each type of disease that emphasize the important the policy for implementation of vaccines on vulnerable societies including children, high-risk societies and mothers.

I am pleased to display for the comparisons of types of disease in different facets within local splines as well as the animation in ggplot for the time line of cases. This would show simultaneously by animation for the decrease of cases in all 9 types of diseases and overall patterns by the local geom_smooth. In both studies of ggplots, the uniqueness of visualization would deliver the pattern of decrease by geom_smooth and facets while the other would deliver for the animation of timeline in disease from 1945 to 2015 for each plots to enjoy the effect of vaccines.

The graph of 9 types of diseases with smoothing splines and its cases conveys the comparisons of cases from 1945 to 2015 with different colors. The original graph is hard to recognize the different quantitative changes of cases but the ggplot with smoothing spline quantitively conveys the change of numbers and its patterns of cases, comparatively. This is more precise and accurate to find from the reader's perspective how the changes occur and see the comparison of 9 different disease. Also, all 9 diseases are conveyed in a different scales where each one of them are shown for its own scales of ranges. This is also easier to recognize for the readers to find the quantitive increase and decrease of the numbers for its own cases. It is also better to estimate also from the reader's point of view for the future timeline as well. 
Also, by the ggplotly, the interactive plot enables for the reader to accurately access the timeline of each plots to find precise information on each number of cases for particular number of years as well as click to zoom in for particular years for visualization, which overwhelms the original plot.

However, the second plot of animation shows the timeline of its graph changes in cases and dynamically attract readers to identify the point that leads each plot. In other words, the readers would identify naturally for the decrease of cases on each diseases to think with, until the animation ends. Also, the line graph with point to animate accurately shows the change of cases for years and dynamically conveys even the small changes in the timeline. The dynamics of timeline for yearly changes for time series data are very well portrayed in the animation of point to observe with. 

extra credit: The most popular and notorious COVID-19 disease for Ontario is portrayed in the https://health-infobase.canada.ca/covid-19/vaccination-coverage/.
As the coverage of vaccine increase over time to be also estimated, the number of cases decreases similar to plots visualize. The vaccine coverage is analogous to plots where interactive plots shows better for the importance of vaccine in a comparative manners. Also, different age groups are shown with different colors to show that there are different result, to be similar to 9 facet graphs I have created.